+++ tests/core/models/test_configs/test_equiformerv2_hydra.yml
@@ -29,7 +29,7 @@ evaluation_metrics:
     misc:
       - energy_forces_within_threshold
   primary_metric: forces_mae
-  
+
 logger:
     name: tensorboard
 
@@ -83,9 +83,6 @@ model:
       module: equiformer_v2_energy_head
     forces:
       module: equiformer_v2_force_head
-    stress:
-      module: rank2_symmetric_head
-      output_name: "stress"
 
 optim:
   batch_size: 5
